requirements.txt
torch
torchvision
mani_skill
gymnasium
wandb
hydra-core
omegaconf
